+++ api_v2/serializers/__init__.py
@@ -22,3 +22,4 @@ from .race import TraitSerializer
 from .race import RaceSerializer
 
 from .creature import CreatureSerializer
+from .creature import CreatureTypeSerializer
