+++ api_v2/views/__init__.py
@@ -1,3 +1,6 @@
 """The initialization for views for open5e's api v2."""
 
 from .views import *
+
+from .creature import CreatureFilterSet, CreatureViewSet
+from .creature import CreatureTypeViewSet
